Adds a new job to alert admins users via email of any new locations.
Locations are considered new when they are undecorated.

The configuration needed to run this job will be done in a separate pull request e.g. cron job configuration and email addresses.

Although there are a reasonably large number of file changes in this PR, a lot are down to a class being renamed.