BAVL-1121 revert location attribute to shared on deletion of last schedule
In room administration, when editing a room that is of type SCHEDULE, it is possible to delete the last schedule item. This results in an error being shown when attempting to re-display the schedules.

If the last schedule is removed, the room should revert back to being of type SHARED, to avoid this error.